move /pages into /src
