+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from 'react';
-import Home from '../src/components/Home';
-import Layout from '../src/components/Layout';
+import Home from '../components/Home';
+import Layout from '../components/Layout';
 
 const HomePage: FunctionComponent = () => {
   const title = 'Wishbone Deli | Home';
